+++ app/plugin/plugin.js
@@ -4,6 +4,7 @@
 const path = require('path');
 const config = require('../../config');
 const VueStore = require('./store');
+const VueRouter = require('./router');
 
 const injectionScriptsPath = config.injectionScripts.path;
 const injectionScripts = config.injectionScripts.files;
@@ -12,6 +13,7 @@ module.exports = class VuePlugin {
     constructor(browser) {
         this._browser = browser;
         this._store = new VueStore(this, browser);
+        this._router = new VueRouter(this, browser);
         this._vueDetected = false;
     }
 
@@ -21,6 +23,10 @@ module.exports = class VuePlugin {
 
     get store() {
         return this._store;
+    }
+
+    get router() {
+        return this._router;
     }