+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.1.0 / ####-##-##
 ==================
 
+  * Vue router get all routes
   * Vue store actions
   * Vue store assertions
   * Detect vue in opened page